requirements.txt
### GENERAL ###
tqdm
numpy
scikit-learn
jupyter

### DEVELOPMENT ###
black
isort
bandit
mypy
pylint

### DATA PROCESSING ###
pillow
seaborn
matplotlib
pandas
pyfonts
retry

### RAG ###
langchain_community
langchain_core
langchain_openai
langchain_huggingface
langchain_chroma
streamlit
chainlit
pypdf
pypdf2
langchain-ollama